[refactor] force users to use wsclient flow ID support
The WithWSClient trait is used to declare a macwire dependency on the
wsclient, but there is really no reason to be using a raw WSClient
anymore. We should always be using the WSClientWithFlow now.
This is going to be a breaking change for all consumers because they
have to instantiate a WSClientWithFlow somewhere now and need at least
a request context in all code paths that result in an HTTP call.